requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==1.59.1
gitpython==3.1.40
jinja2==3.1.2
werkzeug==2.3.7
markdown==3.5.1
colorama==0.4.6
python-dateutil==2.8.2